Base Docker image on "nextcloudci/node:node-7"
"nextcloudci/node" is used in server and some apps to check whether the
JavaScript built files are clean or not. Until recently that Docker
image and "node:lts" had the same NPM version, but the version has been
recently bumped in "node:lts", which causes files built with the
npmbuildbot to be seen as not clean when checked from a
"nextcloudci/node" container. To fix this version mismatch, and given
that we have full control over "nextcloudci/node" but not over
"node:lts", npmbuildbot should use "nextcloudci/node" instead.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>